requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg==3.1.13
django-extensions==3.2.3
django-debug-toolbar==4.3.0
ipython==8.21.0
pgcli==4.0.1
celery==5.3.6
redis==5.0.1
flower==2.0.1
drf-rw-serializers==1.1.1
drf-spectacular==0.27.1
django-environ==0.11.2